Resolve conflicts from upstream
