+++ libraries/DisplayResults.class.php
@@ -625,7 +625,7 @@ class PMA_DisplayResults
             || ($displayParts['sort_lnk'] == '1'))
             && (/*overload*/mb_strlen($db) && !empty($table))
         ) {
-            $the_total   = PMA_Table::countRecords($db, $table);
+            $the_total = $GLOBALS['dbi']->getTable($db, $table)->countRecords();
         }
 
         // if for COUNT query, number of rows returned more than 1
@@ -643,9 +643,10 @@ class PMA_DisplayResults
             // - For a VIEW we (probably) did not count the number of rows
             //   so don't test this number here, it would remove the possibility
             //   of sorting VIEW results.
+            $_table = new PMA_Table($table, $db);
             if (isset($unlim_num_rows)
                 && ($unlim_num_rows < 2)
-                && ! PMA_Table::isView($db, $table)
+                && ! $_table->isView()
             ) {
                 $displayParts['sort_lnk'] = (string) '0';
             }
@@ -1671,7 +1672,8 @@ class PMA_DisplayResults
         }
 
         // generate table create time
-        if (! PMA_Table::isView($this->__get('db'), $this->__get('table'))) {
+        $table = new PMA_Table($this->__get('table'), $this->__get('db'));
+        if (! $table->isView()) {
             $data_html .= '<input class="table_create_time" type="hidden" value="'
                 . PMA_Table::sGetStatusInfo(
                     $this->__get('db'), $this->__get('table'), 'Create_time'
@@ -4649,7 +4651,8 @@ class PMA_DisplayResults
 
         }
 
-        if (PMA_Table::isView($this->__get('db'), $this->__get('table'))
+        $table = new PMA_Table($this->__get('table'), $this->__get('db'));
+        if ($table->isView()
             && ($total == $GLOBALS['cfg']['MaxExactCountViews'])
         ) {
