+++ libraries/tbl_info.inc.php
@@ -57,7 +57,7 @@ if ($showtable) {
     /** @var PMA_String $pmaString */
     $pmaString = $GLOBALS['PMA_String'];
 
-    if (PMA_Table::isView($GLOBALS['db'], $GLOBALS['table'])) {
+    if ($GLOBALS['dbi']->getTable($GLOBALS['db'], $GLOBALS['table'])->isView()) {
         $tbl_is_view     = true;
         $tbl_storage_engine = __('View');
         $show_comment    = null;
@@ -76,9 +76,9 @@ if ($showtable) {
         : $showtable['Collation'];
 
     if (null === $showtable['Rows']) {
-        $showtable['Rows']   = PMA_Table::countRecords(
-            $GLOBALS['db'], $showtable['Name'], true
-        );
+        $showtable['Rows']   = $GLOBALS['dbi']
+            ->getTable($GLOBALS['db'], $showtable['Name'])
+            ->countRecords(true);
     }
     $table_info_num_rows = isset($showtable['Rows']) ? $showtable['Rows'] : 0;
     $row_format = isset($showtable['Row_format']) ? $showtable['Row_format'] : '';
