+++ test/classes/plugin/export/PMA_ExportSql_test.php
@@ -1244,6 +1244,10 @@ class PMA_ExportSql_Test extends PHPUnit_Framework_TestCase
             ->with('res')
             ->will($this->returnValue($row));
 
+        $dbi->expects($this->any())
+            ->method('getTable')
+            ->will($this->returnValue(new PMA_Table('table', 'db')));
+
         $GLOBALS['dbi'] = $dbi;
         $GLOBALS['cfg']['Server']['DisableIS'] = false;
 
@@ -1416,6 +1420,10 @@ class PMA_ExportSql_Test extends PHPUnit_Framework_TestCase
             ->with('res')
             ->will($this->returnValue($row));
 
+        $dbi->expects($this->any())
+            ->method('getTable')
+            ->will($this->returnValue(new PMA_Table('table', 'db')));
+
         $GLOBALS['dbi'] = $dbi;
         $GLOBALS['cfg']['Server']['DisableIS'] = false;
 
@@ -1528,6 +1536,10 @@ class PMA_ExportSql_Test extends PHPUnit_Framework_TestCase
         $dbi->expects($this->once())
             ->method('getError')
             ->will($this->returnValue('error occurred'));
+
+        $dbi->expects($this->any())
+            ->method('getTable')
+            ->will($this->returnValue(new PMA_Table('table', 'db')));
 
         $GLOBALS['dbi'] = $dbi;
         $GLOBALS['cfg']['Server']['DisableIS'] = false;
